fix(semantic-release): rework semantic-release exports again
