+++ packages/semantic-release/src/configs/common/index.ts
@@ -1,0 +1,7 @@
+import SemanticRelease from 'semantic-release'
+import settings from './settings'
+import plugins from './plugins'
+
+const config: Partial<SemanticRelease.GlobalConfig> = {...settings, plugins}
+
+export = config
