+++ packages/semantic-release/src/configs/publish/index.ts
@@ -1,0 +1,7 @@
+import SemanticRelease from 'semantic-release'
+import settings from '../common/settings'
+import plugins from './plugins'
+
+const config: Partial<SemanticRelease.GlobalConfig> = {...settings, plugins}
+
+export = config
