+++ packages/semantic-release/src/exports/common.ts
@@ -1,0 +1,3 @@
+import common from '../configs/common'
+
+export = common
