+++ packages/semantic-release/src/exports/publish.ts
@@ -1,0 +1,3 @@
+import publish from '../configs/publish'
+
+export = publish
